update on LibraryItem Click

+++ app/src/main/java/com/github/lilinsong3/m3player/data/local/entity/PlayList.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class PlayList(
-    @PrimaryKey val songId: Int,
+    @PrimaryKey val songId: Long,
     @ColumnInfo(defaultValue = "(datetime('now'))") val datetime: String
 )
